# openapi/spec.yml
components:
  schemas:
...
    categoryView:
      properties:
        category:
          $ref: '#/components/schemas/categoryView'
        category_id:
          format: uuid
          type: string
        id:
          format: uuid
          type: string
        items:
          items:
            $ref: '#/components/schemas/itemView'
          type: array
        name:
          type: string
      type: object
...
    itemCreate:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      type: object
    itemPartial:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      type: object
...
    itemUpdate:
      properties:
        category_id:
          format: uuid
          type: string
        name:
          type: string
      required:
      - name
      type: object
    itemView:
      properties:
        category:
          $ref: '#/components/schemas/categoryView'
        category_id:
          format: uuid
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
      type: object
...
    loginCreate:
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          type: string
      required:
      - email
      - password
      type: object
    loginPartial:
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          type: string
      type: object
...
    loginUpdate:
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          type: string
      required:
      - email
      - password
      type: object
    loginView:
      properties:
        email:
          format: email
          type: string
        id:
          format: uuid
          type: string
        password:
          format: password
          type: string
      type: object
...
  securitySchemes:
    bearer:
      scheme: bearer
      type: http
    cookieAccessToken:
      in: cookie
      name: x-access-token
      type: apiKey
    cookieRefreshToken:
      in: cookie
      name: x-refresh-token
      type: apiKey
    key:
      in: header
      name: X-API-Key
      type: apiKey
    oauth:
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          scopes:
            admin: Grants access to admin operations
            read: Grants read access
            write: Grants write access
          tokenUrl: https://example.com/oauth/token
      type: oauth2
info:
  description: Invoicing API Reference
  title: Invoicing API
  version: '1.0'
openapi: 3.0.3
paths:
...
  /api/v1/items:
    get:
      description: get all item entities
      operationId: getItems
      parameters:
      - description: Filter Items by name
        in: query
        name: name
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/itemView'
                type: array
          description: successful operation
      summary: get all item entities
      tags:
      - item
    post:
      description: Add a new item entity
      operationId: addItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemCreate'
        description: Create a new item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new item entity
      tags:
      - item
  /api/v1/items/{id}:
    delete:
      description: Delete an existing item entity
      operationId: deleteItem
      parameters:
      - description: item id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid item id
        '404':
          description: item not found
      summary: Delete an existing item entity
      tags:
      - item
    get:
      description: Get an existing item entity
      operationId: getItem
      parameters:
      - description: item id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid item id
        '404':
          description: item not found
      summary: Get an existing item entity
      tags:
      - item
    patch:
      description: Modify an existing item entity (partial update, only provided properties
        will be updated)
      operationId: modifyItem
      parameters:
      - description: item id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemPartial'
        description: Modify an existing item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: item not found
        '405':
          description: Invalid input
      summary: Modify an existing item entity
      tags:
      - item
    put:
      description: Update an existing item entity (all required properties must be
        supplied)
      operationId: updateItem
      parameters:
      - description: item id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/itemUpdate'
        description: Update an existing item entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/itemView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: item not found
        '405':
          description: Invalid input
      summary: Update an existing item entity
      tags:
      - item
  /api/v1/logins:
    get:
      description: get all login entities
      operationId: getLogins
      parameters:
      - description: Filter Logins by email
        in: query
        name: email
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/loginView'
                type: array
          description: successful operation
      summary: get all login entities
      tags:
      - login
    post:
      description: Add a new login entity
      operationId: addLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginCreate'
        description: Create a new login entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginView'
          description: successful operation
        '405':
          description: invalid input
      summary: Add a new login entity
      tags:
      - login
  /api/v1/logins/{id}:
    delete:
      description: Delete an existing login entity
      operationId: deleteLogin
      parameters:
      - description: login id to delete
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginView'
          description: successful operation
        '400':
          description: Invalid login id
        '404':
          description: login not found
      summary: Delete an existing login entity
      tags:
      - login
    get:
      description: Get an existing login entity
      operationId: getLogin
      parameters:
      - description: login id to get
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginView'
          description: successful operation
        '400':
          description: Invalid login id
        '404':
          description: login not found
      summary: Get an existing login entity
      tags:
      - login
    patch:
      description: Modify an existing login entity (partial update, only provided
        properties will be updated)
      operationId: modifyLogin
      parameters:
      - description: login id to partially update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginPartial'
        description: Modify an existing login entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: login not found
        '405':
          description: Invalid input
      summary: Modify an existing login entity
      tags:
      - login
    put:
      description: Update an existing login entity (all required properties must be
        supplied)
      operationId: updateLogin
      parameters:
      - description: login id to update
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginUpdate'
        description: Update an existing login entity
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginView'
          description: successful operation
        '400':
          description: Invalid ID
        '404':
          description: login not found
        '405':
          description: Invalid input
      summary: Update an existing login entity
      tags:
      - login
...
  /login:
    post:
      description: perform a login attempt
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  default: test
                  type: string
                password:
                  default: skgjhsdgklj
                  type: string
        description: Perform a login attempt
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  username:
                    type: string
          description: successful operation
        '400':
          description: bad request
        '403':
          description: invalid username or password
        '405':
          description: invalid input
      summary: Perform a login attempt
security:
- cookieAccessToken: []
  cookieRefreshToken: []
servers:
- url: http://localhost:8001
...